Adicionando métodos relacionados a pesquisa na nova classe Pesquisa
